Deleted copy constructor and assignment operator in classes with explicitly declared destructor.
This is because classes with virtual destructor are used in polymorphism,
and slicing should be avoided.

Classes that use the destructor to release a resource should not have
trivial copy functions to prevent shallow copy of resource handlers,
and double frees.

The most consise way to delete the copy functions is to delete the
move assignment operator.